Fixed RTC bug

+++ include/drivers/rtc/RTC.hpp
@@ -20,7 +20,7 @@ class RTC : public Driver {
         uint8_t hour;
         uint8_t day;
         uint8_t month;
-        uint8_t year;
+        uint32_t year;
         char* getDayNameFromInt(int day);
         int getUpdateInProgress();
         unsigned char getRTCRegister(int reg);
